+++ src/lib/components/methods/nautilus/NautilusMethod.svelte
@@ -3,43 +3,56 @@
 A user interface for the NAUTILUS method.
 -->
 <script lang="ts">
-  //
   import * as _ from "$lib/methods/nautilus/functional_api";
   import { backend } from "$lib/api";
   import type { Backend, Problem } from "$lib/api";
   import { toastStore } from "@skeletonlabs/skeleton";
   import ProblemDetails from "$lib/components/main/ProblemDetails.svelte";
   import GeneralError from "$lib/components/util/undecorated/GeneralError.svelte";
-  import Objective from "$lib/components/main/Objective.svelte";
+  import Objective from "$lib/components/methods/nautilus/Objective.svelte";
   import type {
     ObjectiveData,
     IterationData,
   } from "$lib/methods/nautilus/types";
 
+  /** The problem to be solved. */
   export let problem: Problem;
-
+  /** The method to be used to solve the problem. */
   let method: _.Method;
   $: method = _.nautilus_method(backend, problem);
 
-  let iteration_counter = 0;
+  /**
+   * Flag whether the first iteration has been completed. let
+   * isFirstIterationCompleted = false; $: if (_.is_iterated(method)) {
+   * isFirstIterationCompleted = true; }
+   */
+  /** Flag whether the preferences have been changed. */
+  let isPreferencesChanged = false;
+
+  /** The data for each iteration. */
   export let iterationData: IterationData[] = [];
-
-  let preference_info: number[];
-  let preference_method: number;
-  let n_iterations: number;
-  let short_step: boolean;
-  let step_back: boolean;
-  let use_previous_preference: boolean;
-
-  //preference_info = [20, 20, 10, 10, 40];
-  //preference_method = 2;
-  n_iterations = 5;
-  short_step = false;
-  step_back = false;
-  use_previous_preference = false;
-
+  /** The objectives of the problem. */
   export let objectives: ObjectiveData[] = problem.objectives;
 
+  /** The number of iterations to be performed. */
+  let preference_info: number[];
+  /** The method to be used to rank the objectives. */
+  let preference_method: number;
+  /** The number of iterations to be performed. Initially set to 5. */
+  let n_iterations = 5;
+  /**
+   * Flag whether to perform a short step. Used if preferences are not changed
+   * after step back.
+   */
+  let short_step = false;
+  /** Flag whether to step back to previous iteration. */
+  let step_back = false;
+  /** Use the preference information from previous iteration. */
+  let use_previous_preference = false;
+  /** The iteration counter used to keep track of the number of iterations. */
+  let iteration_counter = 0;
+
+  /** Initialize the method. */
   async function handle_initialize() {
     try {
       method = await _.initialize(method);
@@ -58,6 +71,7 @@ A user interface for the NAUTILUS method.
     }
   }
 
+  /** Function to perform an iteration. */
   async function handle_iterate() {
     if (validate_preferences() === false) {
       return;
@@ -65,16 +79,12 @@ A user interface for the NAUTILUS method.
     if (_.can_iterate(method)) {
       if (_.is_iterated(method) === false) {
         try {
-          console.log("first", preference_info);
-
           method = await _.firstIteraion(
             method,
             n_iterations,
             preference_method,
             preference_info
           );
-          console.log("method", method);
-
           updateIterationSpecs(method);
         } catch (err) {
           toastStore.trigger({
@@ -87,8 +97,6 @@ A user interface for the NAUTILUS method.
         }
       } else {
         try {
-          console.log("consecuent", preference_info);
-
           method = await _.iterate(
             method,
             n_iterations,
@@ -112,6 +120,33 @@ A user interface for the NAUTILUS method.
     }
   }
 
+  /**
+   * Function to go back to the previous iteration. Deletes the last iteration
+   * from iterationData and sets the flags.
+   */
+  async function handleGoBack() {
+    // Prevent going back beyond the first iteration
+    if (iterationData.length > 1) {
+      iterationData.pop(); // Remove the last iteration
+
+      short_step = isPreferencesChanged ? false : true;
+      use_previous_preference = isPreferencesChanged ? true : false;
+      step_back = true;
+    } else {
+      toastStore.trigger({
+        message: "Cannot go back further.",
+        background: "variant-filled-warning",
+        timeout: 5000,
+      });
+    }
+  }
+
+  /**
+   * Function to validate the preferences entered by the user.
+   *
+   * @returns True if the preferences are valid, false otherwise and triggers
+   *   correct error.
+   */
   function validate_preferences() {
     if (preference_method === 1) {
       let hasPositiveRank = preference_info.some((element) => element > 0);
@@ -141,7 +176,6 @@ A user interface for the NAUTILUS method.
       });
       if (sum !== 100) {
         toastStore.trigger({
-          // prettier-ignore
           message: "Please enter weights that sum up to 100.",
           background: "variant-filled-error",
           timeout: 5000,
@@ -152,23 +186,46 @@ A user interface for the NAUTILUS method.
     return true;
   }
 
+  /**
+   * Function to handle the rank update event.
+   *
+   * @param event
+   */
   function handleRankUpdate(event: {
     detail: { index: number; value: number };
   }) {
     const { index, value } = event.detail;
     preference_info[index] = value;
+    isPreferencesChanged = true;
   }
 
+  /**
+   * Function to handle the toggle rank weight switch.
+   *
+   * @param event
+   */
   function handleToggleRankWeight(event: { detail: { useRank: boolean } }) {
     const { useRank } = event.detail;
     preference_method = useRank ? 1 : 2;
+    isPreferencesChanged = true;
   }
 
+  /**
+   * Function to update the number of iterations.
+   *
+   * @param event
+   */
   function updateIterations(event: Event) {
     const inputElement = event.target as HTMLInputElement;
     n_iterations = inputElement ? parseInt(inputElement.value, 10) : 1;
+    isPreferencesChanged = true;
   }
 
+  /**
+   * Function to update the iteration data.
+   *
+   * @param method
+   */
   function updateIterationSpecs(method: {
     _tag?: "Initialized" | "Iterated";
     backend?: Backend;
@@ -196,6 +253,10 @@ A user interface for the NAUTILUS method.
     iterationData = [...iterationData, newIteration];
 
     iteration_counter++;
+    isPreferencesChanged = false;
+    short_step = false;
+    use_previous_preference = false;
+    step_back = false;
   }
 </script>
 
@@ -229,16 +290,20 @@ A user interface for the NAUTILUS method.
         <button class="btn variant-filled" on:click={handle_iterate}
           >Iterate</button
         >
+        {#if iterationData.length > 1}
+          <button
+            class="btn variant-filled"
+            on:click={handleGoBack}
+            disabled={isPreferencesChanged}>Go Back</button
+          >
+        {/if}
       </div>
     {:else}
       <GeneralError />
     {/if}
   </div>
 </div>
-
 {#if _.is_initialized(method) || _.is_iterated(method)}
-  {console.log("iterations", iterationData)}
-
   <div class="objectives-container">
     <Objective
       {iterationData}
@@ -249,30 +314,12 @@ A user interface for the NAUTILUS method.
   </div>
 {/if}
 
-<!-- Current solution section -->
-
-<!-- Buttons section -->
-<div class="actions">
-  <button class="primary">YES</button>
-  <button class="secondary">NO</button>
-  <!-- ... other buttons ... -->
-</div>
-
 <style>
   .objectives-container {
     display: flex;
     overflow-x: auto;
   }
 
-  .primary {
-    background-color: #4a90e2;
-    color: white;
-    /* ... other styles ... */
-  }
-  .secondary {
-    background-color: #f5f5f5;
-    /* ... other styles ... */
-  }
   .iteration-controls {
     display: flex;
     align-items: center;
@@ -285,11 +332,11 @@ A user interface for the NAUTILUS method.
   }
 
   .iteration-input {
-    flex: 0 1 auto; /* Allow the input to grow and shrink as needed but not fill the space */
+    flex: 0 1 auto;
     padding: 8px;
     border: 1px solid #ccc;
     border-radius: 4px;
     box-shadow: inset 0 1px 3px rgba(0, 0, 0, 0.1);
-    margin-right: 10px; /* Ensure some space between the input and the button */
+    margin-right: 10px;
   }
 </style>
